Move PendingChangesListener to deployment.internal
This seems to be the place where it
belongs.
