+++ subprojects/core/src/main/java/org/gradle/internal/session/BuildSessionScopeServices.java
@@ -42,6 +42,7 @@ import org.gradle.cache.internal.InMemoryCacheDecoratorFactory;
 import org.gradle.cache.internal.scopes.DefaultBuildTreeScopedCache;
 import org.gradle.cache.scopes.BuildTreeScopedCache;
 import org.gradle.deployment.internal.DefaultDeploymentRegistry;
+import org.gradle.deployment.internal.PendingChangesManager;
 import org.gradle.groovy.scripts.internal.DefaultScriptSourceHasher;
 import org.gradle.groovy.scripts.internal.ScriptSourceHasher;
 import org.gradle.initialization.BuildCancellationToken;
@@ -59,7 +60,6 @@ import org.gradle.internal.classpath.ClassPath;
 import org.gradle.internal.event.DefaultListenerManager;
 import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.file.Deleter;
-import org.gradle.internal.filewatch.PendingChangesManager;
 import org.gradle.internal.hash.ChecksumService;
 import org.gradle.internal.hash.DefaultChecksumService;
 import org.gradle.internal.isolation.IsolatableFactory;
