+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/ContinuousBuildActionExecutor.java
@@ -20,6 +20,8 @@ import org.gradle.api.logging.LogLevel;
 import org.gradle.deployment.internal.Deployment;
 import org.gradle.deployment.internal.DeploymentInternal;
 import org.gradle.deployment.internal.DeploymentRegistryInternal;
+import org.gradle.deployment.internal.PendingChangesListener;
+import org.gradle.deployment.internal.SingleFirePendingChangesListener;
 import org.gradle.execution.CancellableOperationManager;
 import org.gradle.execution.DefaultCancellableOperationManager;
 import org.gradle.execution.PassThruCancellableOperationManager;
@@ -35,8 +37,6 @@ import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.execution.WorkInputListener;
 import org.gradle.internal.execution.WorkInputListeners;
 import org.gradle.internal.file.Stat;
-import org.gradle.internal.filewatch.PendingChangesListener;
-import org.gradle.internal.filewatch.SingleFirePendingChangesListener;
 import org.gradle.internal.invocation.BuildAction;
 import org.gradle.internal.logging.text.StyledTextOutput;
 import org.gradle.internal.logging.text.StyledTextOutputFactory;
